requirements.txt
PyYAML~=5.3.1
wxPython~=4.0.7.post2